cleaned code

+++ pytorch-a2c-ppo-acktr/algo/ppo.py
@@ -63,7 +63,7 @@ class PPO():
                 action_loss = -torch.min(surr1, surr2).mean()
 
                 value_loss = F.mse_loss(return_batch, values)
-                psc_add = torch.tensor(psc_add, requires_grad=True)
+                psc_add = torch.tensor(psc_add, requires_grad=True, dtype = torch.float)
 
                 self.optimizer.zero_grad()
                 (value_loss * self.value_loss_coef + action_loss -
